requirements.txt
Django==3.1.2
djangorestframework==3.12.1
drf-yasg==1.17.1
django-heroku==0.3.1
gunicorn==20.0.4
PyJWT==1.7.1
django-cors-headers==3.5.0